Run in a browser: HTML shell, byte-swapped data, JSPI yields
The port now runs end-to-end in Chromium: data loads, the levelset is found,
the title screen and main menu appear, and keyboard input advances them.
Verified by driving headless Chromium over the DevTools protocol - pressing
Enter at the title screen reaches doMenu("Main menu:", 4) and paints 268 text
pixels, with no exceptions.

Endianness, offline (tools/mkdata.py, replacing mkdata.sh)
  The converter is now struct-aware. Only the structures the game memcpy's or
  casts a pointer at are swapped:
    MLST  levelsetdata          MTXT  gametextdata
    MLEV  levelfiledata + map_data + one leveldata per level
    MBG   bg_filedata + one bg_data per background
  GFX (ma_tiles / ma_sprts) is pixel data and is passed through untouched -
  the graphics ports will read those rows big-endian. Level payloads are left
  alone too: level.c reads enemies, flying platforms and triggers one unsigned
  char at a time, so nothing in them is endian-sensitive.

  Caught while writing this: ma_bckgr is tagged MBG, not GFX. An earlier draft
  keyed bg_filedata off GFX and so byte-swapped 44 bytes of sprite and tile
  pixel data. There is now a regression check that ma_tiles.bin is identical
  to its raw container slice.

levelfiledata.Total_size: int -> short
  TIGCC compiles with 16-bit int (-mshort). As a 32-bit field every level
  offset shifted and the last one fell outside the file; as 16-bit all eight
  offsets land in range and decode to sensible dimensions (111x14, 95x12,
  120x10). This is the only int field in any struct in the codebase.

Busy-wait deadlock
  WaitKeyPress/WaitKeyReleased spun on _rowread without yielding. That is fine
  on hardware but pegs the browser's main thread, so no key event can ever be
  delivered and the loop cannot exit - the renderer sat at 90% CPU. They now
  call browser_yield(), a JSPI await of requestAnimationFrame, so callers stay
  synchronous. Also fixed _rowread itself: its argument is a row MASK where a
  zero bit selects a row, not a row index, so _rowread(0) means "any key".

compat/library.js and library_pre.js are replaced by EM_ASYNC_JS in
compat/gray.c and compat/kbd.c; get_menu_string died with the text-system
revert, so no JS files remain. gray.c now renders the two planes to RGBA and
hands that to the canvas, which was previously wired up only half way.

shell.html is a standalone page with the canvas and a browser-to-key-matrix
map. Known gap: F_4x6 is not in the AMS image and falls back to the 6x8 face,
which is wider, so the statusbar fields overlap - it lays out at hardcoded x
positions assuming 4px glyphs.

+++ Mario/level.h
@@ -265,7 +265,9 @@ typedef struct{
 
 typedef struct{
 	short Nr_of_levels;
-	int Total_size;
+	// TIGCC compiles with 16-bit int (-mshort); this field is 2 bytes in the
+	// data files, so it must stay 16-bit here or every level offset shifts.
+	short Total_size;
 	
 	unsigned short Levels[20];//holds the offset from start of file to start of the levelstruct
